requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
bitarray==2.7.5
blinker==1.4
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
command-not-found==0.3
constantly==15.1.0
cryptography==41.0.1
cytoolz==0.12.1
daphne==4.0.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==2.1.0
Django==4.2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-heroku==0.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
frozenlist==1.3.3
gunicorn==21.2.0
hexbytes==0.3.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
jeepney==0.7.1
jsonschema==4.17.3
keyring==23.5.0
kombu==5.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lru-dict==1.2.0
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
openai==0.27.8
packaging==23.1
parsimonious==0.9.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==5.4.1
redis==5.0.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
schedule==1.2.0
SecretStorage==3.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
systemd-python==234
texttable==1.6.4
toolz==0.12.0
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.3
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
web3==6.5.0
websocket-client==1.2.3
websockets==11.0.3
whitenoise==6.5.0
yarl==1.9.2
zipp==1.0.0
zope.interface==6.0
